Correctly make BasicErrorResponse a DTO
